rename wn1.c wm1.h to wm1_hotswap.c, wm1_hotswap.h
